requirements.txt
Flask==2.3.3
google-cloud-bigquery==3.20.0
#functions-framework==3.5.0
pyarrow>= 3.0.0
